+++ data/src/file_gridfs_plugin.py
@@ -93,3 +93,6 @@ class FileGridFSPlugin(colony.Plugin):
 
     def exists(self, connection, file_name):
         return self.system.exists(connection, file_name)
+
+    def exists_directory(self, connection, directory_name):
+        return self.system.exists_directory(connection, directory_name)